requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
beautifulsoup4==4.9.3
boto3==1.18.15
botocore==1.21.15
certifi==2021.5.30
cfn-flip==1.2.3
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-email-verification==0.1.0
django-import-export==2.5.0
django-livereload-server==0.3.2
django-ratelimit==3.0.1
durationpy==0.5
et-xmlfile==1.1.0
future==0.18.2
hjson==3.0.2
idna==3.2
Jinja2==3.0.1
jmespath==0.10.0
kappa==0.6.0
MarkupPy==1.14
MarkupSafe==2.0.1
mysqlclient==2.0.3
numpy==1.21.2
odfpy==1.4.1
openpyxl==3.0.7
pandas==1.3.2
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tornado==6.1
tqdm==4.62.0
troposphere==3.0.2
urllib3==1.26.6
Werkzeug==0.16.1
whitenoise==5.3.0
wsgi-request-logger==0.4.6
xlrd==2.0.1
xlwt==1.3.0
zappa==0.53.0
